requirements.txt
aiohttp==3.11.18
motor==3.7.1
pydantic==2.11.4
pydantic_settings==2.9.1
Pyrogram==2.0.106
TgCrypto==1.2.5
python-dotenv==1.1.0
qrcode==8.2
Pillow==11.1.0
Requests==2.32.3
